+++ berkeley-function-call-leaderboard/bfcl/model_handler/handler_map.py
@@ -19,6 +19,7 @@ from bfcl.model_handler.local_inference.bielik import BielikHandler
 from bfcl.model_handler.local_inference.deepseek import DeepseekHandler
 from bfcl.model_handler.local_inference.deepseek_coder import DeepseekCoderHandler
 from bfcl.model_handler.local_inference.deepseek_reasoning import DeepseekReasoningHandler
+from bfcl.model_handler.local_inference.falcon_fc import Falcon3FCHandler
 from bfcl.model_handler.local_inference.gemma import GemmaHandler
 from bfcl.model_handler.local_inference.glaive import GlaiveHandler
 from bfcl.model_handler.local_inference.glm import GLMHandler
@@ -33,8 +34,8 @@ from bfcl.model_handler.local_inference.mistral_fc import MistralFCHandler
 from bfcl.model_handler.local_inference.phi import PhiHandler
 from bfcl.model_handler.local_inference.quick_testing_oss import QuickTestingOSSHandler
 from bfcl.model_handler.local_inference.qwen import QwenHandler
+from bfcl.model_handler.local_inference.qwen_fc import QwenFCHandler
 from bfcl.model_handler.local_inference.salesforce import SalesforceHandler
-from bfcl.model_handler.local_inference.falcon_fc import Falcon3FCHandler
 
 # TODO: Add meta-llama/Llama-3.1-405B-Instruct
 
@@ -150,12 +151,19 @@ local_inference_handler_map = {
     "THUDM/glm-4-9b-chat": GLMHandler,
     "Qwen/Qwen2-1.5B-Instruct": QwenHandler,
     "Qwen/Qwen2-7B-Instruct": QwenHandler,
+    "Qwen/Qwen2.5-0.5B-Instruct-FC": QwenFCHandler,
     "Qwen/Qwen2.5-0.5B-Instruct": QwenHandler,
+    "Qwen/Qwen2.5-1.5B-Instruct-FC": QwenFCHandler,
     "Qwen/Qwen2.5-1.5B-Instruct": QwenHandler,
+    "Qwen/Qwen2.5-3B-Instruct-FC": QwenFCHandler,
     "Qwen/Qwen2.5-3B-Instruct": QwenHandler,
+    "Qwen/Qwen2.5-7B-Instruct-FC": QwenFCHandler,
     "Qwen/Qwen2.5-7B-Instruct": QwenHandler,
+    "Qwen/Qwen2.5-14B-Instruct-FC": QwenFCHandler,
     "Qwen/Qwen2.5-14B-Instruct": QwenHandler,
+    "Qwen/Qwen2.5-32B-Instruct-FC": QwenFCHandler,
     "Qwen/Qwen2.5-32B-Instruct": QwenHandler,
+    "Qwen/Qwen2.5-72B-Instruct-FC": QwenFCHandler,
     "Qwen/Qwen2.5-72B-Instruct": QwenHandler,
     "Team-ACE/ToolACE-8B": LlamaHandler,
     "openbmb/MiniCPM3-4B": MiniCPMHandler,